requirements.txt
dash
plotly
PyYAML
uvloop
asyncpg
pandas
dash-bootstrap-components
psycopg2-binary